+++ sdk/src/shared.rs
@@ -2,14 +2,13 @@ use std::collections::hash_map::DefaultHasher;
 use std::hash::{Hash, Hasher};
 use std::num::NonZeroUsize;
 use std::sync::Arc;
-use std::sync::atomic::{AtomicBool, Ordering};
 
 use lru::LruCache;
 use nostr::EventId;
-use nostr::signer::AsyncNostrSigner;
 use nostr_database::NostrDatabase;
 use tokio::sync::Mutex;
 
+use crate::authenticator::Authenticator;
 use crate::monitor::Monitor;
 use crate::policy::AdmitPolicy;
 use crate::transport::websocket::WebSocketTransport;
@@ -22,10 +21,9 @@ const MAX_VERIFICATION_CACHE_SIZE: usize = 128_000;
 pub(crate) struct SharedState {
     pub(crate) database: Arc<dyn NostrDatabase>,
     pub(crate) transport: Arc<dyn WebSocketTransport>,
-    signer: Option<Arc<dyn AsyncNostrSigner>>,
-    nip42_auto_authentication: Arc<AtomicBool>,
     verification_cache: Arc<Mutex<LruCache<u64, ()>>>,
     pub(crate) admit_policy: Option<Arc<dyn AdmitPolicy>>,
+    pub(crate) authenticator: Option<Arc<dyn Authenticator>>,
     pub(crate) monitor: Option<Monitor>,
 }
 
@@ -33,9 +31,8 @@ impl SharedState {
     pub(crate) fn new(
         database: Arc<dyn NostrDatabase>,
         transport: Arc<dyn WebSocketTransport>,
-        signer: Option<Arc<dyn AsyncNostrSigner>>,
         admit_policy: Option<Arc<dyn AdmitPolicy>>,
-        nip42_auto_authentication: bool,
+        authenticator: Option<Arc<dyn Authenticator>>,
         monitor: Option<Monitor>,
     ) -> Self {
         let max_verification_cache_size: NonZeroUsize =
@@ -45,22 +42,11 @@ impl SharedState {
         Self {
             database,
             transport,
-            signer,
-            nip42_auto_authentication: Arc::new(AtomicBool::new(nip42_auto_authentication)),
             verification_cache: Arc::new(Mutex::new(LruCache::new(max_verification_cache_size))),
             admit_policy,
+            authenticator,
             monitor,
         }
-    }
-
-    #[inline]
-    pub(crate) fn is_auto_authentication_enabled(&self) -> bool {
-        self.nip42_auto_authentication.load(Ordering::SeqCst)
-    }
-
-    pub(crate) fn automatic_authentication(&self, enable: bool) {
-        self.nip42_auto_authentication
-            .store(enable, Ordering::SeqCst);
     }
 
     #[inline]
@@ -68,12 +54,9 @@ impl SharedState {
         &self.database
     }
 
-    pub(crate) fn has_signer(&self) -> bool {
-        self.signer.is_some()
-    }
-
-    pub(crate) fn signer(&self) -> Option<&Arc<dyn AsyncNostrSigner>> {
-        self.signer.as_ref()
+    #[inline]
+    pub(crate) fn is_authenticator_available(&self) -> bool {
+        self.authenticator.is_some()
     }
 
     pub(crate) async fn verified(&self, id: &EventId) -> bool {